Refactor repo to use src format, remove setup.py
