+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2022 MarBL
+Copyright (c) 2024 MarBL
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
