+++ config/config.json
@@ -1,6 +1,4 @@
 {
-    "static": true,
-    "alpha": 0.3,
     "identity": 90,
     "sparsity": 10,
     "palette": "OrRd_7",
@@ -15,7 +13,7 @@
         100
     ],
     "output_dir": "Chr1_cen_plots",
-    "input": [
-        "./test/Chr1_cen.fa"
+    "fasta": [
+        "./sequences/Chr1_cen.fa"
     ]
 }
