--- moddotplot/__main__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python3
-import sys
-from moddotplot.estimate_identity import *
-from moddotplot.moddotplot import main
-from moddotplot.parse_fasta import *
-
-sys.exit(main())
